Revert "changed file structure to be more organized"
This reverts commit 450050891a9b0b40a7064f86e15c24040623104d.
