+++ client/src/components/Reviews-Ratings/BreakdownComp.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 // import { sumHelper } from './helpers';
-import helpers from '../../../helpPlease';
+import helpers from './helpers';
 
 function BreakdownComp({filters, updateFilters, reviewsMeta}) {
   // data comes from reviewsMeta
