+++ client/src/components/Reviews-Ratings/Sorting.jsx
@@ -1,8 +1,5 @@
 import React, { useState } from 'react';
-import helpers from '../../../helpPlease';
-const sortHelpfulness = helpers.sortHelpfulness;
-const sortNewest = helpers.sortNewest;
-const sortRelevance = helpers.sortRelevance;
+import { sortHelpfulness, sortNewest, sortRelevance } from './helpers';
 
 function Sorting({ sortedFilteredReviews, updateSetSortedFilteredReviews }) {
   // this is a single function that sorts the array in three different ways depending on what user selects in the dropdown. It references functions in helpers.js file
